+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ENHANCEMENTS
 
 * Bootstrap Yorc with a Vault instance ([GH-282](https://github.com/ystia/yorc/issues/282))
+* Refactor Slurm jobs ([GH-220](https://github.com/ystia/yorc/issues/220))
 
 ### FEATURES